requirements.txt
atomicwrites==1.2.1
attrs==18.2.0
certifi==2018.10.15
chardet==3.0.4
Click==7.0
clickclick==1.2.2
colorama==0.4.0
connexion==1.5.3
Flask==1.0.2
Flask-Caching==1.4.0
idna==2.7
inflection==0.3.1
itsdangerous==0.24
Jinja2==2.10
jsonschema==2.6.0
MarkupSafe==1.0
more-itertools==4.3.0
pluggy==0.8.0
py==1.7.0
pytest==3.9.2
python-dateutil==2.7.4
python-dotenv==0.9.1
PyYAML==3.13
requests==2.20.0
six==1.11.0
swagger-spec-validator==2.4.1
urllib3==1.24
watson-developer-cloud==2.2.6
websocket-client==0.47.0
Werkzeug==0.14.1
